Read response from backend to prevend TCP RSTs
If the response isn't read, then a proper FIN/ACK message won't
be returned to the backend and the connection will eventually
get reset. Also causes entries in backend firewall logs as the
RST is being sent to a closed connection.

Changes the URL from ...:6544/ to ...:6544/Myth/GetHostName because
the response is roughly 2900 bytes smaller.