References #215. Adds "Connection:", "Close" to prevent TCP RSTs
Also, reads the response from the backend. This allows the
conversation to make an orderly close (otherwise it has unread
data and doesn't issue a FIN-ACK.)  Gzipping was turned off just
for these very short responses (<String>host</String>.)

Finally, a 5 second timeout was added. This may be up for 
debate (what if the backend is busy, as opposed to gone?)