retest #219
I added some catches in for the MythServicesApiRuntimeException.

Unfortunately, i think the only thing we can do at this point is allow
it to fail gracefully.  If the incoming data isn't completed or
unparsable in any way, I don't think it can partially proceed.  These
parsing exceptions are at too deep a level to allow the app to try to
access any of the data and load it.