Fixes #234 From beta user on MythTV -users channel.
Allows canceling an attempt to connect to the backend. Note,
user elected not to use the issues list. And, I can't duplicate
this problem (I have a tablet with WiFi only, no 4G or data plan
like the user.) Cancel appears, but the dialog is closed so fast
that I can't test the new Cancel button.